Add the FAKE_LIGHT_CURVE config option, and tweak the light levels.
Using the LIGHT_CURVE option brings in a lot of floating-point math
functions, which take about 50k of flash.  As an alternative, the
FAKE_LIGHT_CURVE option approximates a 0.33 light curve but with much
less flash space requirements.

This commit also changes the light level thresholds to make them more
in line with typical light levels encountered in the world.

The sensor will give numbers from 0-980, but >100 generally means
there's a light source in the field of view; that should pretty much
be "totally constricted" for iris adjustments.  In some quick
experiments, I get:

* Dark room w/ laptop and TV: 0
* Dimly-lit room: 0-2
* Brightly-lit room: 5-10
* Dark room, under a desk lamp: 31-33
* Brightly-lit room, under a desk lamp: 35-45
* Flashlight visible in FOV: 110-120 (regardless of ambient light)
* Flashlight shining on board: 980

Clearly, light curve correction is important here!  Also, the range
0-980 isn't terribly useful.  Using the ranges in this commit will
open the iris 3/4 of the way when the light level is about 2.5,
halfway when the light level is about 5, and close the iris 1/4 of the
way when the light is about 20.  It will be at its maximum open and
close at 0 and 48, respectively.

You can get the fake light curve table here with Mathematica on your
Raspberry Pi, by entering:
    Piecewise[Table[{Rescale[x, {2^n, 2^(n - 1)}, {2^n, 2^(n - 1)}^(1/3)],
      x <= 2^n}, {n, -4, 0}]] /. x <= 1 -> True // N
Then press Shift-Enter.  For this commit, I also changed the smallest
term with the output of "Rescale[x, {2^-4, 0}, {2^-4, 0}^(1/3)] // N",
to fix up the section from 0 to 1/8, which is a critical area.