Add accelerometer-based eye tracking.
This commit adds the ability for the eye to track vertically based on
the tilt of the board.  If the board is tilted forwards, the eye looks
up; if the board is tilted backwards, the eye looks down.  In other
words, the eye looks forwards regardless of the tilt of the board.

It currently is based around the Hallowing board, where the X axis
indicates the pitch rotation of the board; other configurations can be
added without difficulty.

I haven't yet worked out good values for the default eye; this
commit's values currently are tuned for the dragon eye.